[data] Initial sample data set.
This has yet to be tokenized, as this is how it was provided by Arshad. The
formatted file will need to be reformatteed for easier processing. Not sure how
I will need the word frequency yet, as I may throw it away in favor of a Perl
Storable data file as per the experimental code my my SNLP class.