[reorg] Slight reorganization to pave the way for more algorithms.
Basically, each algorithm shall implement the empty roll
AI::Subjectivity::Seeder, which requires the build() method. This will pave the
way for other algorithms, such as MSOL, PSL, TTL, SWN, GI, or a combination of
any of these algorithms. I'm keeping these classes as plain Moose classes/roles
for now to avoid the penalty of MooseX::* (or so I've been told). Will revisit
this when the need arises.