[oo] Restructuring needed by the new score.pl.
This restructuring move methods and members from Seed.pm to the module
that uses them, thus improving encapsalation. Also, The Seed classes
no longer need the SeedArgs object, as that was a hack to save time.
Instead, this now passes the required parameters in to the functions.
As a side effect, a new read_data_files method had to be added to
each class (no pure virtuals in Moose?) that is responsible of
loading its own data.

Lastly, the seed.pl was restructured so the prior lexicons can always
be loaded independently of the data files. For that to happen, the
members in AI::Subjectivity::Seed had to be distinguished so that
the lexicon data is only the lexicon data.